use uint256 for x,y,size

+++ src/Land/erc721/LandBaseToken.sol
@@ -78,7 +78,7 @@ contract LandBaseToken is ERC721BaseToken {
      * @param y The top left y coordinate of the new quad
      * @param data extra data to pass to the transfer
      */
-    function mintQuad(address to, uint16 size, uint16 x, uint16 y, bytes calldata data) external {
+    function mintQuad(address to, uint256 size, uint256 x, uint256 y, bytes calldata data) external {
         require(to != address(0), "to is zero address");
         require(
             isMinter(msg.sender),
@@ -114,8 +114,8 @@ contract LandBaseToken is ERC721BaseToken {
                 "Already minted as 12x12"
             );
         } else {
-            for (uint16 x12i = x; x12i < toX; x12i += 12) {
-                for (uint16 y12i = y; y12i < toY; y12i += 12) {
+            for (uint256 x12i = x; x12i < toX; x12i += 12) {
+                for (uint256 y12i = y; y12i < toY; y12i += 12) {
                     uint256 id12x12 = LAYER_12x12 + x12i + y12i * GRID_SIZE;
                     require(_owners[id12x12] == 0, "Already minted as 12x12");
                 }
@@ -125,8 +125,8 @@ contract LandBaseToken is ERC721BaseToken {
         if (size <= 6) {
             require(_owners[LAYER_6x6 + (x/6) * 6 + ((y/6) * 6) * GRID_SIZE] == 0, "Already minted as 6x6");
         } else {
-            for (uint16 x6i = x; x6i < toX; x6i += 6) {
-                for (uint16 y6i = y; y6i < toY; y6i += 6) {
+            for (uint256 x6i = x; x6i < toX; x6i += 6) {
+                for (uint256 y6i = y; y6i < toY; y6i += 6) {
                     uint256 id6x6 = LAYER_6x6 + x6i + y6i * GRID_SIZE;
                     require(_owners[id6x6] == 0, "Already minted as 6x6");
                 }
@@ -136,8 +136,8 @@ contract LandBaseToken is ERC721BaseToken {
         if (size <= 3) {
             require(_owners[LAYER_3x3 + (x/3) * 3 + ((y/3) * 3) * GRID_SIZE] == 0, "Already minted as 3x3");
         } else {
-            for (uint16 x3i = x; x3i < toX; x3i += 3) {
-                for (uint16 y3i = y; y3i < toY; y3i += 3) {
+            for (uint256 x3i = x; x3i < toX; x3i += 3) {
+                for (uint256 y3i = y; y3i < toY; y3i += 3) {
                     uint256 id3x3 = LAYER_3x3 + x3i + y3i * GRID_SIZE;
                     require(_owners[id3x3] == 0, "Already minted as 3x3");
                 }
@@ -156,7 +156,7 @@ contract LandBaseToken is ERC721BaseToken {
         _checkBatchReceiverAcceptQuad(msg.sender, address(0), to, size, x, y, data);
     }
 
-    function _idInPath(uint256 i, uint16 size, uint16 x, uint16 y) internal pure returns(uint256) {
+    function _idInPath(uint256 i, uint256 size, uint256 x, uint256 y) internal pure returns(uint256) {
         uint256 row = i / size;
         if(row % 2 == 0) { // alow ids to follow a path in a quad
             return (x + (i%size)) + ((y + row) * GRID_SIZE);
@@ -172,7 +172,7 @@ contract LandBaseToken is ERC721BaseToken {
     /// @param x The top left x coordinate of the quad
     /// @param y The top left y coordinate of the quad
     /// @param data additional data
-    function transferQuad(address from, address to, uint16 size, uint16 x, uint16 y, bytes calldata data) external {
+    function transferQuad(address from, address to, uint256 size, uint256 x, uint256 y, bytes calldata data) external {
         require(from != address(0), "from is zero address");
         require(to != address(0), "can't send to zero address");
         bool metaTx = msg.sender != from && _metaTransactionContracts[msg.sender];
@@ -194,9 +194,9 @@ contract LandBaseToken is ERC721BaseToken {
         address operator,
         address from,
         address to,
-        uint16 size,
-        uint16 x,
-        uint16 y,
+        uint256 size,
+        uint256 x,
+        uint256 y,
         bytes memory data
     ) internal {
         if (to.isContract() && _checkInterfaceWith10000Gas(to, ERC721_MANDATORY_RECEIVER)) {
@@ -221,9 +221,9 @@ contract LandBaseToken is ERC721BaseToken {
     function batchTransferQuad(
         address from,
         address to,
-        uint16[] calldata sizes,
-        uint16[] calldata xs,
-        uint16[] calldata ys,
+        uint256[] calldata sizes,
+        uint256[] calldata xs,
+        uint256[] calldata ys,
         bytes calldata data
     ) external {
         require(from != address(0), "from is zero address");
@@ -239,7 +239,7 @@ contract LandBaseToken is ERC721BaseToken {
         }
         uint256 numTokensTransfered = 0;
         for (uint256 i = 0; i < sizes.length; i++) {
-            uint16 size = sizes[i];
+            uint256 size = sizes[i];
             _transferQuad(from, to, size, xs[i], ys[i]);
             numTokensTransfered += size * size;
         }
@@ -250,7 +250,7 @@ contract LandBaseToken is ERC721BaseToken {
             uint256[] memory ids = new uint256[](numTokensTransfered);
             uint256 counter = 0;
             for (uint256 j = 0; j < sizes.length; j++) {
-                uint16 size = sizes[j];
+                uint256 size = sizes[j];
                 for (uint256 i = 0; i < size*size; i++) {
                     ids[counter] = _idInPath(i, size, xs[j], ys[j]);
                     counter++;
@@ -263,7 +263,7 @@ contract LandBaseToken is ERC721BaseToken {
         }
     }
 
-    function _transferQuad(address from, address to, uint16 size, uint16 x, uint16 y) internal {
+    function _transferQuad(address from, address to, uint256 size, uint256 x, uint256 y) internal {
         if (size == 1) {
             uint256 id1x1 = x + y * GRID_SIZE;
             address owner = _ownerOf(id1x1);
@@ -288,7 +288,7 @@ contract LandBaseToken is ERC721BaseToken {
         return false;
     }
 
-    function _regroup(address from, address to, uint16 size, uint16 x, uint16 y) internal {
+    function _regroup(address from, address to, uint256 size, uint256 x, uint256 y) internal {
         require(x % size == 0 && y % size == 0, "Invalid coordinates");
         require(x <= GRID_SIZE - size && y <= GRID_SIZE - size, "Out of bounds");
 
@@ -305,12 +305,12 @@ contract LandBaseToken is ERC721BaseToken {
         }
     }
 
-    function _regroup3x3(address from, address to, uint16 x, uint16 y, bool set) internal returns (bool) {
+    function _regroup3x3(address from, address to, uint256 x, uint256 y, bool set) internal returns (bool) {
         uint256 id = x + y * GRID_SIZE;
         uint256 quadId = LAYER_3x3 + id;
         bool ownerOfAll = true;
-        for (uint16 xi = x; xi < x+3; xi++) {
-            for (uint16 yi = y; yi < y+3; yi++) {
+        for (uint256 xi = x; xi < x+3; xi++) {
+            for (uint256 yi = y; yi < y+3; yi++) {
                 ownerOfAll = _checkAndClear(from, xi + yi * GRID_SIZE) && ownerOfAll;
             }
         }
@@ -329,12 +329,12 @@ contract LandBaseToken is ERC721BaseToken {
         }
         return ownerOfAll;
     }
-    function _regroup6x6(address from, address to, uint16 x, uint16 y, bool set) internal returns (bool) {
+    function _regroup6x6(address from, address to, uint256 x, uint256 y, bool set) internal returns (bool) {
         uint256 id = x + y * GRID_SIZE;
         uint256 quadId = LAYER_6x6 + id;
         bool ownerOfAll = true;
-        for (uint16 xi = x; xi < x+6; xi += 3) {
-            for (uint16 yi = y; yi < y+6; yi += 3) {
+        for (uint256 xi = x; xi < x+6; xi += 3) {
+            for (uint256 yi = y; yi < y+6; yi += 3) {
                 bool ownAllIndividual = _regroup3x3(from, to, xi, yi, false);
                 uint256 id3x3 = LAYER_3x3 + xi + yi * GRID_SIZE;
                 uint256 owner3x3 = _owners[id3x3];
@@ -361,12 +361,12 @@ contract LandBaseToken is ERC721BaseToken {
         }
         return ownerOfAll;
     }
-    function _regroup12x12(address from, address to, uint16 x, uint16 y, bool set) internal returns (bool) {
+    function _regroup12x12(address from, address to, uint256 x, uint256 y, bool set) internal returns (bool) {
         uint256 id = x + y * GRID_SIZE;
         uint256 quadId = LAYER_12x12 + id;
         bool ownerOfAll = true;
-        for (uint16 xi = x; xi < x+12; xi += 6) {
-            for (uint16 yi = y; yi < y+12; yi += 6) {
+        for (uint256 xi = x; xi < x+12; xi += 6) {
+            for (uint256 yi = y; yi < y+12; yi += 6) {
                 bool ownAllIndividual = _regroup6x6(from, to, xi, yi, false);
                 uint256 id6x6 = LAYER_6x6 + xi + yi * GRID_SIZE;
                 uint256 owner6x6 = _owners[id6x6];
@@ -392,12 +392,12 @@ contract LandBaseToken is ERC721BaseToken {
         }
         return ownerOfAll;
     }
-    function _regroup24x24(address from, address to, uint16 x, uint16 y, bool set) internal returns (bool) {
+    function _regroup24x24(address from, address to, uint256 x, uint256 y, bool set) internal returns (bool) {
         uint256 id = x + y * GRID_SIZE;
         uint256 quadId = LAYER_24x24 + id;
         bool ownerOfAll = true;
-        for (uint16 xi = x; xi < x+24; xi += 12) {
-            for (uint16 yi = y; yi < y+24; yi += 12) {
+        for (uint256 xi = x; xi < x+24; xi += 12) {
+            for (uint256 yi = y; yi < y+24; yi += 12) {
                 bool ownAllIndividual = _regroup12x12(from, to, xi, yi, false);
                 uint256 id12x12 = LAYER_12x12 + xi + yi * GRID_SIZE;
                 uint256 owner12x12 = _owners[id12x12];
